decimal field added to appointment checkout table

+++ Appointment/models.py
@@ -229,7 +229,7 @@ class AppointmentCheckout(models.Model):
     service_price = models.PositiveIntegerField(default=0, null=True, blank=True)
     total_price = models.PositiveIntegerField(default=0, null=True, blank=True)
     
-    service_commission = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)    
+    service_commission = models.DecimalField(default = 0 , null=True, blank=True, decimal_places=5, max_digits=8)    
     service_commission_type = models.CharField( max_length=50 , default = '')
